Follow Peter Donald lead:
"
I just made all the dependencies on jakarta-ant include the inherited runtime.
This is to facilitate ant having dependencies in the future without ant
"users" having such dependencies.

My question is, should the xerces project be included as a runtime library of
ant. It is not strictly a dependency because any jaxp parser will do. In
theory another parser could be substituted but in practice it does not seem
to be the case.

Adding it as a runtime dependency would have the following effects.

* Many projects depend on xerces for the sole reason of running ant. If the
project does not actually include any xml dependencies they can instead just
depend on ant and its runtime
* Projects that depend on ant but use a different parser (is there any of
these?) would have tonot depend on ant and runtime but on ant individually
and the parser individually. In effect there is no change for these projects.
"


git-svn-id: http://svn.apache.org/repos/asf/jakarta/cactus/trunk@236472 13f79535-47bb-0310-9956-ffa450edef68