Adding support for server-side XSLT transformations to the ServletTestRunner. This code currently uses reflection to avoid a direct dependancy on the TraX APIs at runtime. It is assumed that the TraX APIs will be available at compile time, as they are bundled with Ant.
To enable the server-side transformations, an XSLT processor needs to be
on the test web-apps classpath (JDK1.4 will do, too). Second, a new
initialization parameter named 'xsl-stylesheet' needs to be added to the
servlets definition in web.xml. Third, the HTTP request parameter
'transform' needs to be present in the request to the servlet, regardless
of the parameters value.

I'll need to document this stuff and probably do some polishing. Would be
cool to get some feedback on whether this works for people.


git-svn-id: http://svn.apache.org/repos/asf/jakarta/cactus/trunk@236866 13f79535-47bb-0310-9956-ffa450edef68