First cut of a <webxmlmerge> task, with very much the same goals as Vincent's <webmerge> task. The good thing about this variant is that it requires much more code to offer less features ;-)
But seriously: this task merges servlet and filter definitions from two
web deployment descriptors into a single descriptor. It takes the servlet
version into account, and will not add the same servlet/filter twice. It
can also merge the init-params of a servlet or filter that already exists
in the original descriptor, although uniqueness of parameters is
currently not checked.

The major increase in lines of code is because the task is more intelligent
about the descriptor, i.e. it knows the concepts of filters and servlets.
However, I really need to cut the code amount because there's currently
much duplication between the handling of filters and servlets.

This task uses the Ant <xmlcatalog> type to provide offline capabilities,
which I believe to be cleaner than the NoopResolver. It does not yet
merge stuff like <login-config> and <security-constraint>. I could go on,
but...


git-svn-id: http://svn.apache.org/repos/asf/jakarta/cactus/trunk@237409 13f79535-47bb-0310-9956-ffa450edef68