CPAN for Cactus... well, not quite ;-)
Dependencies are listed as (project,artifact,version) in dependencies.xml.
This file is not intended to be modified by the user. The download script
is (you guessed it) download.xml. This build file requires Ant 1.6
because of its use of <macrodef>.

There are still problems: it's hard to override which versions to download.
More importantly, the file to download is specified by dependencies.xml,
while the target (local) file name is specified by build.properties. Thus,
we need to pay attention that both are in sync.

Anyway, this script has helped me a lot in quickly getting Cactus to
build on "virgin" platforms.


git-svn-id: http://svn.apache.org/repos/asf/jakarta/cactus/trunk@238737 13f79535-47bb-0310-9956-ffa450edef68