Corrections for sample projects
Adjust building of the samples modules ("with-samples" profile).
Fix some download URLs (Tomcat and JBoss).
Fix some uses of explicit version numbers, where none are needed
(stabelize build).

TODO jetty does still not work as the classpath to the tools.jar is not
correctly handeled.
```
Unable to find a javac compiler;
com.sun.tools.javac.Main is not on the classpath.
Perhaps JAVA_HOME does not point to the JDK.
```